Add notifications for transferring gSOL
