+++ packages/app/src/common/utils.ts
@@ -2,6 +2,10 @@ import { LAMPORTS_PER_SOL, type PublicKey } from "@solana/web3.js";
 import { type SignerWalletAdapterProps } from "@solana/wallet-adapter-base";
 import { MAX_NUM_PRECISION } from "@sunrisestake/client";
 import BN from "bn.js";
+import {
+  NotificationType,
+  notifyTransaction,
+} from "./components/notifications";
 
 const ZERO = new BN(0);
 
@@ -77,6 +81,15 @@ function debounce<F extends (...args: Parameters<F>) => ReturnType<F>>(
   };
 }
 
+const handleError = (error: Error): void => {
+  notifyTransaction({
+    type: NotificationType.error,
+    message: "Transaction failed",
+    description: error.message,
+  });
+  console.error(error);
+};
+
 export {
   ZERO,
   CARBON_PRICE_USD_CENTS_PER_TONNE,
@@ -89,4 +102,5 @@ export {
   toBN,
   toFixedWithPrecision,
   walletIsConnected,
+  handleError,
 };
